started poking fun - timeline can now play the main audioplayer, and to beat.
next steps: visualize beats, make a way to add beats. A good way to do this might be with our good friend scrubrect. if we can make it into a tool, we might be able to instantiate and modify them at runtime, as well as modify settings like divisions. scale setting and expandrect setting would help (more divisions, etc)